CRUD de cursos terminado *Muestra error 500 en update, pero igualmente actualiza.

+++ src/main/java/com/udea/Fabrica_g6_v2/repositories/MateriaRepository.java
@@ -5,6 +5,8 @@ import com.udea.Fabrica_g6_v2.models.Materia;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface MateriaRepository  extends JpaRepository<Materia,Long> {
+public interface MateriaRepository  extends JpaRepository<Materia,Integer> {
 }
